did a lot of reorganizing
