+++ nln/Info/Maps/CharacterDescription.js
