+++ nln/Inventory/Maps/Weapons.js
