+++ src/sass/assets/temp-background.jpg
