+++ src/sass/pages/_home.scss
@@ -1,40 +1,3 @@
-.header-bar {
-  &__profile-button {
-    position: fixed;
-    right: 0;
-    top: 0;
-    padding: 3rem 4rem;
-    &--static {
-      position: absolute;
-    }
-  }
-  &__profile-img {
-    border-radius: 100%;
-    width: 5rem;
-    cursor: pointer;
-    box-shadow: 0 0 1.5rem $color-grey-light-2;
-    &:hover {
-      filter: brightness(0.6);
-    }
-  }
-}
-
-.header {
-  background-image: url(../../assets/temp-background.jpg);
-  background-size: cover;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-bottom: 1px solid $color-grey-light-1;
-  &__logotype {
-    font-family: "Audiowide", cursive;
-    text-transform: uppercase;
-    letter-spacing: 0.4rem;
-    font-size: 6rem;
-    color: $color-grey-dark-2;
-  }
-}
-
 .container {
   width: 80%;
   margin: 4rem auto;
